+++ clickhouse-admin/test-utils/Cargo.toml
@@ -8,4 +8,6 @@ license = "MPL-2.0"
 workspace = true
 
 [dependencies]
+camino.workspace = true
 clickward.workspace = true
+dropshot.workspace = true
