+++ clickhouse-admin/test-utils/src/lib.rs
@@ -4,7 +4,10 @@
 
 //! Integration testing facilities for clickhouse-admin
 
-use clickward::{BasePorts, Deployment};
+use camino::Utf8PathBuf;
+use clickward::{BasePorts, Deployment, DeploymentConfig};
+use dropshot::test_util::{log_prefix_for_test, LogContext};
+use dropshot::{ConfigLogging, ConfigLoggingLevel};
 
 pub const DEFAULT_CLICKHOUSE_ADMIN_BASE_PORTS: BasePorts = BasePorts {
     keeper: 29000,
@@ -14,6 +17,28 @@ pub const DEFAULT_CLICKHOUSE_ADMIN_BASE_PORTS: BasePorts = BasePorts {
     clickhouse_interserver_http: 29400,
 };
 
-//pub fn default_clickhouse_cluster_test_deployment() -> Deployment {
-//
-//}
+pub fn default_clickhouse_cluster_test_deployment(
+    path: Utf8PathBuf,
+) -> Deployment {
+    let base_ports = DEFAULT_CLICKHOUSE_ADMIN_BASE_PORTS;
+
+    let config = DeploymentConfig {
+        path,
+        base_ports,
+        cluster_name: "oximeter_cluster".to_string(),
+    };
+
+    Deployment::new(config)
+}
+
+pub fn default_clickhouse_log_ctx_and_path() -> (LogContext, Utf8PathBuf) {
+    let logctx = LogContext::new(
+        "clickhouse_cluster",
+        &ConfigLogging::StderrTerminal { level: ConfigLoggingLevel::Info },
+    );
+
+    let (parent_dir, _prefix) = log_prefix_for_test("clickhouse_cluster");
+    let path = parent_dir.join("clickward_test");
+
+    (logctx, path)
+}
